MySQL連携およびAPI実装

+++ src/main/kotlin/com/azul/spring/project/controller/UserController.kt
@@ -1,6 +1,6 @@
 package com.azul.spring.project.controller
 
-import com.azul.spring.project.entity.User
+import com.azul.spring.project.domain.User
 import com.azul.spring.project.service.UserService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.ResponseEntity
@@ -16,7 +16,7 @@ class UserController @Autowired constructor(private val userService: UserService
     }
     
     @RequestMapping("/api/users/{id}")
-    fun api_show(@PathVariable("id") id: Long): ResponseEntity<User> {
+    fun api_show(@PathVariable("id") id: Int): ResponseEntity<User> {
         return ResponseEntity.ok(userService.findById(id))
     }
 
@@ -27,11 +27,11 @@ class UserController @Autowired constructor(private val userService: UserService
 
     @RequestMapping("/api/users", method = arrayOf(RequestMethod.POST))
     fun api_create(@RequestBody user: User): ResponseEntity<User> {
-        return ResponseEntity.ok(userService.save(User( user.id, user.name, user.age)))
+        return ResponseEntity.ok(userService.save(User(user.id, user.name, user.age)))
     }
 
     @RequestMapping("/api/users/{id}", method = arrayOf(RequestMethod.DELETE))
-    fun api_destroy(id: Long): ResponseEntity<User> {
+    fun api_destroy(id: Int): ResponseEntity<User> {
         var user = userService.findById(id)
         userService.destroy(user)
         return ResponseEntity.ok(user)
